+++ src/main/java/com/oncha/oncha_web/feature/BaseRestController.java
@@ -1,0 +1,8 @@
+package com.oncha.oncha_web.feature;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@RequestMapping("/api")
+public abstract class BaseRestController {
+
+}
