+++ src/main/java/com/oncha/oncha_web/feature/category/SubscribeController.java
@@ -1,0 +1,8 @@
+package com.oncha.oncha_web.feature.category.subscribe;
+
+import org.springframework.web.bind.annotation.GetMapping;
+
+public class SubscribeController {
+    @GetMapping("/subscribe")
+    public String Subscribe(){return "/category/subscribe";}
+}
