+++ src/main/java/com/oncha/oncha_web/feature/category/company/controller/CompanyController.java
@@ -1,0 +1,4 @@
+package com.oncha.oncha_web.feature.category.company.controller;
+
+public class CompanyController {
+}
