--- src/main/java/com/oncha/oncha_web/feature/company/controller/CompanyController.java
@@ -1,4 +1,0 @@
-package com.oncha.oncha_web.feature.company.controller;
-
-public class CompanyController {
-}
