+++ src/main/java/com/oncha/oncha_web/feature/product/productReview/controller/ProductReviewController.java
@@ -1,0 +1,4 @@
+package com.oncha.oncha_web.feature.product.productReview.controller;
+
+public class ProductReviewController {
+}
