+++ src/main/java/com/oncha/oncha_web/feature/product/productReview/service/ProductReviewService.java
@@ -1,0 +1,4 @@
+package com.oncha.oncha_web.feature.product.productReview.service;
+
+public class ProductReviewService {
+}
