--- src/main/java/com/oncha/oncha_web/feature/productReview/controller/ProductReviewController.java
@@ -1,4 +1,0 @@
-package com.oncha.oncha_web.feature.productReview.controller;
-
-public class ProductReviewController {
-}
