--- src/main/java/com/oncha/oncha_web/feature/productReview/service/ProductReviewService.java
@@ -1,4 +1,0 @@
-package com.oncha.oncha_web.feature.productReview.service;
-
-public class ProductReviewService {
-}
