--- src/main/java/com/oncha/oncha_web/feature/subscribe/SubscribeController.java
@@ -1,8 +1,0 @@
-package com.oncha.oncha_web.feature.subscribe;
-
-import org.springframework.web.bind.annotation.GetMapping;
-
-public class SubscribeController {
-    @GetMapping("/subscribe")
-    public String Subscribe(){return "/category/subscribe";}
-}
