--- src/main/java/com/oncha/oncha_web/feature/tea_finder/TeaFinderController.java
@@ -1,9 +1,0 @@
-package com.oncha.oncha_web.feature.tea_finder;
-
-import org.springframework.web.bind.annotation.GetMapping;
-
-public class TeaFinderController {
-    @GetMapping("/finder")
-    public String TeaFinder(){return "category/tea_finder";}
-
-}
